backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic-settings>=2.0
python-multipart>=0.0.9
ultralytics
pillow>=10.0
